Link to docs (#333)
* increase maximum error offset for CI tests

We try to detect how long durations took. However there is no nice way
to time this in JSDOM. Instead we take snapshots every
requestAnimationFrame and when things change we also write down the
time.

This solution is not ideal and results in false positives (especially on
CI environments).

However, it is good enough to ensure that the duration is not 0 and not
500.

* cleanup README's and link to docs site

* remove readme's in favor of doc site

This will be easier, so that we don't have to maintain multiple repo's.